impl

+++ Assets/_Chi/Scripts/Statistics/NpcStats.cs
@@ -10,7 +10,7 @@ namespace _Chi.Scripts.Statistics
     {
         // not using Stat because this class will exist in many thousands instances
         
-        public int speed = 1;
+        public float speed = 1;
         public int rotationSpeed = 500;
     
         public float contactDamage;
